# Improve GraphQL::Dataloader::Source#sync's efficiency