@glasses618 2023-06-18
Sorry, I found I don't need to dup the `pending_keys` which `sync` was called. Let me make some changes.