@rmosolgo 2023-06-19
Thanks for this contribution, it looks great! Out of curiousity, did you update the benchmarks above after removing the `pending_keys.dup` call? I wonder if that would improve the memory usage of the Dataloader profile...