@glasses618 2023-06-20
> Thanks for this contribution, it looks great! Out of curiousity, did you update the benchmarks above after removing the `pending_keys.dup` call? I wonder if that would improve the memory usage of the Dataloader profile...

Yes, the current benchmark report represents the latest version. The original version of the benchmark, which utilized `pending_keys = @pending_keys.dup` (https://github.com/rmosolgo/graphql-ruby/commit/35542e3b8c997f52329f90dce9d678add456a3ff), can be summarized as follows:

In Ruby 2.7.8, it the increases ~0.94% memory and ~2.8% objects. In Ruby 3.0.6, it increases ~0.96% memory and ~2.5% objects.

After observing the memory and objects increases in the original version,  I notice that `pending_keys = @pending_keys.dup` can be removed.